Code refactoring. Structure modification. Add enable and config mode to base CliHandlerImpl
