+++ cloudshell/networking/devices/runners/interfaces/firmware_operations_interface.py
