+++ cloudshell/tests/operations/test_configuration_operations.py
@@ -3,7 +3,7 @@ from unittest import TestCase
 import jsonpickle
 from mock import MagicMock
 
-from cloudshell.networking.devices.operations.configuration_operations import ConfigurationOperations
+from cloudshell.networking.devices.runners.configuration_operations import ConfigurationOperations
 from cloudshell.shell.core.context import ResourceCommandContext, ReservationContextDetails, ResourceContextDetails
 from cloudshell.shell.core.interfaces.save_restore import OrchestrationSavedArtifact
